refactor structured output route
